refactor: centralize socket inbound types in core for shared handler
Define TranscribeMessage, StreamAmbientMessage, and RecordingSocketInboundMessage
in core so RecordingButtonBase can narrow the WebSocket switch while dictation and
ambient components still declare their mode-specific message generic.
